* 'develop' of https://github.com/jenyglee/final-project:
  Feat : 공지사항 백엔드 조건문 수정
  Feat : 공지사항 단건조회 수정중
  Feat : JwtFilter CORS에러 PATCH 추가
  Feat : 수정중
  Feat : JwtFilter CORS에러 수정
  Update .gitignore
  Feat : 코스추천 단건 조회 수정 완료
  Feat : 관리자페이지 코스추천 전체조회 완료
  오타수정

# Conflicts:
#	back/src/main/java/com/project/sparta/recommendCourse/entity/RecommendCourseBoard.java

+++ back/src/main/java/com/project/sparta/noticeBoard/service/NoticeBoardServiceImpl.java
@@ -8,6 +8,8 @@ import com.project.sparta.noticeBoard.entity.NoticeBoard;
 import com.project.sparta.noticeBoard.entity.NoticeCategoryEnum;
 import com.project.sparta.noticeBoard.repository.NoticeBoardRepository;
 import com.project.sparta.user.entity.User;
+import com.project.sparta.user.entity.UserRoleEnum;
+import com.project.sparta.user.entity.UserRoleEnum.Authority;
 import com.querydsl.core.QueryResults;
 import java.time.LocalDateTime;
 import java.util.stream.Collectors;
@@ -34,26 +36,31 @@ public class NoticeBoardServiceImpl implements NoticeBoardService {
     @Override
     public void createNoticeBoard(NoticeBoardRequestDto requestDto, User user) {
         // 0: SERVICE, 1: UPDATE, 2: EVENT
-        NoticeBoard noticeBoard = new NoticeBoard(user, requestDto.getTitle(),
-            requestDto.getContents(), requestDto.getCategory());
-        noticeBoardRepository.saveAndFlush(noticeBoard);
+        if(user.getRole()== UserRoleEnum.ADMIN) {
+            NoticeBoard noticeBoard = new NoticeBoard(user, requestDto.getTitle(),
+                requestDto.getContents(), requestDto.getCategory());
+            noticeBoardRepository.saveAndFlush(noticeBoard);
+        }
     }
 
     //공지글 삭제
     @Override
     public void deleteNoticeBoard(Long id, User user) {
-        NoticeBoard noticeBoard = noticeBoardRepository.findByIdAndUser_NickName(id,
-            user.getNickName()).orElseThrow(() -> new CustomException(NOT_FOUND_POST));
-        noticeBoardRepository.delete(noticeBoard);
+        if(user.getRole()== UserRoleEnum.ADMIN) {
+            NoticeBoard noticeBoard = noticeBoardRepository.findById(id)
+                .orElseThrow(() -> new CustomException(NOT_FOUND_POST));
+            noticeBoardRepository.delete(noticeBoard);
+        }
     }
 
     //공지글 수정
     @Override
     public void updateNoticeBoard(Long id, NoticeBoardRequestDto requestDto, User user) {
-        NoticeBoard noticeBoard = noticeBoardRepository.findByIdAndUser_NickName(id,
-            user.getNickName()).orElseThrow(() -> new CustomException(NOT_FOUND_POST));
-        noticeBoard.update(requestDto.getTitle(), requestDto.getContents(),
-            requestDto.getCategory());
+        if(user.getRole()== UserRoleEnum.ADMIN)
+        {
+            NoticeBoard noticeBoard = noticeBoardRepository.findById(id).orElseThrow(() -> new CustomException(NOT_FOUND_POST));
+            noticeBoard.update(requestDto.getTitle(), requestDto.getContents(), requestDto.getCategory());
+        }
     }
 
     //공지글 단건 조회
@@ -61,18 +68,18 @@ public class NoticeBoardServiceImpl implements NoticeBoardService {
     public NoticeBoardResponseDto getNoticeBoard(Long id, User user) {
 
         NoticeBoard noticeBoard = noticeBoardRepository.findById(id)
-            .orElseThrow(() -> new CustomException(NOT_FOUND_POST));
+                .orElseThrow(() -> new CustomException(NOT_FOUND_POST));
 
         return new NoticeBoardResponseDto(noticeBoard.getId(), noticeBoard.getUser().getNickName(),
-            noticeBoard.getTitle(),
-            noticeBoard.getContents(), noticeBoard.getCategory(), noticeBoard.getModifiedAt(),
-            noticeBoard.getCreateAt());
+                noticeBoard.getTitle(),
+                noticeBoard.getContents(), noticeBoard.getCategory(), noticeBoard.getModifiedAt(),
+                noticeBoard.getCreateAt());
     }
 
     //공지글 전체조회
     @Override
     public PageResponseDto<List<NoticeBoardResponseDto>> getAllNoticeBoard(int page, int size,
-        String category, User user) {
+                                                                           String category, User user) {
         long total;
         List<NoticeBoardResponseDto> content = new ArrayList<>();
         if (!category.isEmpty()) {
@@ -89,7 +96,7 @@ public class NoticeBoardServiceImpl implements NoticeBoardService {
             // 1. 페이징으로 요청해서 조회
             //        PageRequest pageRequest = PageRequest.of(page, size);
             QueryResults<NoticeBoardResponseDto> noticeBoardList = noticeBoardRepository.findAllByCategory(
-                categoryEnum, page, size);
+                    categoryEnum, page, size);
 
             // 2. 데이터, 전체 개수 추출
 //            Page<NoticeBoardResponseDto> allList = results.map(noticeBoard -> new NoticeBoardResponseDto(noticeBoard.getId(), noticeBoard.getUser().getNickName(), noticeBoard.getTitle(),
@@ -108,17 +115,17 @@ public class NoticeBoardServiceImpl implements NoticeBoardService {
             //        .map(CommunityBoardOneResponseDto::new)
             //        .collect(Collectors.toList());
             content = boardList.getContent()
-                .stream()
-                .map((obj) -> new NoticeBoardResponseDto(
-                    obj.getId(),
-                    obj.getUser().getNickName(),
-                    obj.getTitle(),
-                    obj.getContents(),
-                    obj.getCategory(),
-                    obj.getCreateAt(),
-                    obj.getModifiedAt()
-                ))
-                .collect(Collectors.toList());
+                    .stream()
+                    .map((obj) -> new NoticeBoardResponseDto(
+                            obj.getId(),
+                            obj.getUser().getNickName(),
+                            obj.getTitle(),
+                            obj.getContents(),
+                            obj.getCategory(),
+                            obj.getCreateAt(),
+                            obj.getModifiedAt()
+                    ))
+                    .collect(Collectors.toList());
 //            Long id, String username, String title, String contents, NoticeCategoryEnum category, LocalDateTime
 //            createdAt, LocalDateTime modifiedAt
         }
